group commands into one cog, refactoring
